removing the .DS-Store file
This is Mac OS specific file, not needed in repository.

DS_Store stands for Desktop Services Store and it holds meta-information about your folder's thumbnails, settings, etc. in macOS operating system. These files are created any time you navigate to a file or folder from the Finder on a Mac.